Rename einops.py to ein_ops.py
This change avoids naming conflicts with the popular third-party einops package.
The rename ensures better clarity and prevents unintended import collisions in
environments where both are present.
