+++ axlearn/common/ssm_kernels/ssd_kernels.py
@@ -33,7 +33,7 @@ from jax import lax
 from jax._src.lax.control_flow import for_loop
 from jax.experimental import pallas as pl
 
-from axlearn.common.einops import rearrange, repeat
+from axlearn.common.ein_ops import rearrange, repeat
 from axlearn.common.utils import Tensor
